+++ sdc/rabbit/__init__.py
@@ -11,4 +11,4 @@ all = [
     MessageConsumer,
 ]
 
-__version__ = '0.3.0'
+__version__ = '0.3.1'