+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Unreleased
+
+### 0.3.1
 - Add tornado to requirements file
 - Pass tx_id in call to process method